move first few widgets to uni_common_ui
