+++ packages/uni_app/lib/view/bus_stop_selection/widgets/bus_stop_selection_row.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:uni/model/entities/bus_stop.dart';
 import 'package:uni/model/providers/lazy/bus_stop_provider.dart';
-import 'package:uni/view/common_widgets/row_container.dart';
+import 'package:uni_common_ui/common/row_container.dart';
 
 class BusStopSelectionRow extends StatefulWidget {
   const BusStopSelectionRow(this.stopCode, this.stopData, {super.key});
