+++ packages/uni_app/lib/view/course_unit_info/widgets/course_unit_files_row.dart
@@ -7,8 +7,8 @@ import 'package:uni/controller/local_storage/file_offline_storage.dart';
 import 'package:uni/generated/l10n.dart';
 import 'package:uni/model/entities/course_units/course_unit_file.dart';
 import 'package:uni/model/providers/startup/session_provider.dart';
-import 'package:uni/view/common_widgets/pulse_animation.dart';
-import 'package:uni/view/common_widgets/toast_message.dart';
+import 'package:uni_common_ui/common/pulse_animation.dart';
+import 'package:uni_common_ui/common/toast_message.dart';
 
 class CourseUnitFilesRow extends StatefulWidget {
   const CourseUnitFilesRow(this.file, {super.key});
