+++ packages/uni_app/lib/view/home/widgets/next_exams_card.dart
@@ -3,9 +3,9 @@ import 'package:provider/provider.dart';
 import 'package:uni/model/entities/exam.dart';
 import 'package:uni/utils/date_time_formatter.dart';
 import 'package:uni/view/common_widgets/date_rectangle.dart';
-import 'package:uni/view/common_widgets/row_container.dart';
 import 'package:uni/view/exams/widgets/exam_row.dart';
 import 'package:uni/view/locale_notifier.dart';
+import 'package:uni_common_ui/common/row_container.dart';
 
 class NextExamsWidget extends StatelessWidget {
   const NextExamsWidget({required this.exams, super.key});
