+++ packages/uni_common_ui/lib/common/pulse_animation.dart
